+++ src/api/users/index.ts
@@ -1,9 +1,8 @@
 import { Router } from "express";
-import { getAllUsersController, createUserController } from "./user.controller";
+import { getAllUsersController } from "./user.controller";
 
 const router = Router();
 
 router.get("/", getAllUsersController);
-router.post("/", createUserController);
 
 export default router;
