+++ src/api/users/user.service.ts
@@ -6,11 +6,3 @@ export const getUsers = () => {
   return prisma.user.findMany();
 };
 
-export const createUser = (input: any) => {
-  return prisma.user.create({
-    data: {
-      email: input.email,
-      password: input.password,
-    },
-  });
-};
