+++ src/auth/auth.controller.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from "express";
 const bcrypt = require("bcrypt");
 const jwt = require("jsonwebtoken");
-import { findUser } from "./auth.service";
+import { findUser, createUser } from "./auth.service";
 
 export const loginUserController = async (req: Request, res: Response) => {
   try {
@@ -28,5 +28,31 @@ export const loginUserController = async (req: Request, res: Response) => {
       status: "fail",
       message: error.message,
     });
+  }
+};
+
+export const signUpUserController = async (req: Request, res: Response) => {
+  try {
+    const passwordRegex =
+      /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[!@#$%^&*]).{8,}$/;
+    const isValidPassword = passwordRegex.test(req.body.password);
+
+    if (!isValidPassword) {
+      return res
+        .status(400)
+        .json({ message: "The password is not strong enough" });
+    }
+
+    const user = {
+      email: req.body.email,
+      password: await bcrypt.hash(req.body.password, 12),
+    };
+
+    const userCreated = await createUser(user);
+    res
+      .status(201)
+      .json({ message: "User created successfully", data: userCreated });
+  } catch (error: any) {
+    res.status(500).json({ message: error.message });
   }
 };
